@@ -8,36 +8,38 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.15",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/fire": "^5.1.3",
-    "@angular/forms": "^7.2.15",
-    "@angular/http": "^7.2.15",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/pwa": "^0.13.9",
-    "@angular/router": "^7.2.15",
-    "@angular/service-worker": "^7.2.15",
-    "@ngxs/storage-plugin": "^3.4.3",
-    "@ngxs/store": "^3.4.3",
-    "core-js": "^2.6.8",
-    "firebase": "^6.1.0",
+    "@angular/animations": "~8.1.3",
+    "@angular/cdk": "^8.1.1",
+    "@angular/common": "~8.1.3",
+    "@angular/compiler": "~8.1.3",
+    "@angular/core": "~8.1.3",
+    "@angular/fire": "^5.2.1",
+    "@angular/forms": "~8.1.3",
+    "@angular/material": "^8.1.1",
+    "@angular/platform-browser": "~8.1.3",
+    "@angular/platform-browser-dynamic": "~8.1.3",
+    "@angular/pwa": "^0.801.2",
+    "@angular/router": "~8.1.3",
+    "@angular/service-worker": "^8.1.3",
+    "@ngxs/storage-plugin": "^3.5.0",
+    "@ngxs/store": "^3.5.0",
+    "firebase": "^6.3.3",
     "hammerjs": "^2.0.8",
     "rxjs": "^6.5.2",
-    "zone.js": "^0.8.29"
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular/cdk": "^7.3.7",
-    "@angular/cli": "^8.0.0",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "^7.2.15",
-    "@types/node": "^12.0.3",
-    "codelyzer": "^5.0.1",
-    "tslint": "^5.16.0",
-    "typescript": "^3.2.4"
+    "@angular-devkit/architect": "^0.801.2",
+    "@angular-devkit/build-angular": "^0.801.2",
+    "@angular/cli": "^8.1.2",
+    "@angular/compiler-cli": "~8.1.3",
+    "@angular/language-service": "~8.1.3",
+    "@types/node": "^12.6.8",
+    "codelyzer": "^5.1.0",
+    "fuzzy": "^0.1.3",
+    "inquirer": "^6.5.0",
+    "inquirer-autocomplete-prompt": "^1.0.1",
+    "tslint": "^5.18.0",
+    "typescript": "^3.4.5"
   }
 }
