@@ -1,9 +1,9 @@
 import * as firebase from 'firebase/app';
 import { BehaviorSubject } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { map, take } from 'rxjs/operators';
 
 import { AngularFireAuth } from '@angular/fire/auth';
-import { State, StateContext, Action, NgxsOnInit, Store } from '@ngxs/store';
+import { State, StateContext, Action, Store } from '@ngxs/store';
 import {
 	IAuthState, CheckSession, LoginSuccess, LoginFailed, LogoutSuccess, LoginIdentity, Logout, AuthToken,
 	LoginEmail, LoginLink, AuthInfo, LoginToken, LoginSetup, LoginCredential, LoginAnon, AuthOther
@@ -22,8 +22,8 @@ import { IRegister } from '@dbase/data/data.schema';
 import { addWhere } from '@dbase/fire/fire.library';
 import { IQuery } from '@dbase/fire/fire.interface';
 
-import { getLocalStore, delLocalStore, prompt } from '@lib/window.library';
-import { getPath } from '@lib/object.library';
+import { getLocalStore, delLocalStore, prompt } from '@lib/browser.library';
+import { getPath, cloneObj } from '@lib/object.library';
 import { isNull } from '@lib/type.library';
 import { dbg } from '@lib/logger.library';
 
@@ -37,14 +37,12 @@ import { dbg } from '@lib/logger.library';
 		current: null,																	// the effective User (used in 'impersonate' mode)
 	}
 })
-export class AuthState implements NgxsOnInit {
+export class AuthState {
 	private dbg = dbg(this);
 	private user: firebase.User | null = null;
 	private _memberSubject = new BehaviorSubject(null as firebase.auth.AdditionalUserInfo | null);
 
 	constructor(private afAuth: AngularFireAuth, private sync: SyncService, private store: Store, private snack: SnackService, private navigate: NavigateService) { this.init(); }
-
-	ngxsOnInit(ctx: StateContext<IAuthState>) { this.init(); }
 
 	private init() {
 		this.dbg('init');
@@ -53,13 +51,21 @@ export class AuthState implements NgxsOnInit {
 	}
 
 	get memberSubject() {
+		if (this._memberSubject.isStopped)							// start a new Subject
+			this._memberSubject = new BehaviorSubject(null as firebase.auth.AdditionalUserInfo | null);
 		return this._memberSubject;
+	}
+
+	get auth() {
+		return this.afAuth.authState
+			.pipe(take(1))
+			.toPromise();
 	}
 
 	private async authSuccess(ctx: StateContext<IAuthState>, user: firebase.User | null, credential: firebase.auth.AuthCredential | null) {
 		if (user) {
 			if (credential) {														// have we been redirected here, via credential?
-				const response = await user.linkAndRetrieveDataWithCredential(credential);
+				const response = await user.linkWithCredential(credential);
 				ctx.patchState({ info: response.additionalUserInfo });
 			}
 		}
@@ -68,9 +74,12 @@ export class AuthState implements NgxsOnInit {
 
 	/** Commands */
 	@Action(CheckSession)														// check Authentication status
-	checkSession(ctx: StateContext<IAuthState>, { user }: CheckSession) {
-		const state = ctx.getState();
+	private checkSession(ctx: StateContext<IAuthState>, { user }: CheckSession) {
 		this.dbg('%s', user ? `${user.displayName} is logged in` : 'not logged in');
+
+		if (isNull(user) && isNull(this.user)) {
+			this.navigate.route(ROUTE.login);						// redirect to LoginComponent
+		}
 
 		if (isNull(user) && !isNull(this.user)) {			// TODO: does this affect OAuth logins
 			ctx.dispatch(new Logout());									// User logged-out
@@ -89,7 +98,7 @@ export class AuthState implements NgxsOnInit {
 	 * Or, an anonymous user can link a Facebook account and then, later, sign in with Facebook to continue using our App.
 	 */
 	@Action(LoginCredential)													// attempt to link multiple providers
-	async loginCredential(ctx: StateContext<IAuthState>, { link }: LoginCredential) {
+	private async loginCredential(ctx: StateContext<IAuthState>, { link }: LoginCredential) {
 		const methods = await this.afAuth.auth.fetchSignInMethodsForEmail(link.email);
 
 		switch (methods[0]) {														// check the first-method
@@ -115,7 +124,7 @@ export class AuthState implements NgxsOnInit {
 
 	/** Attempt to sign-in a User via authentication by a federated identity provider */
 	@Action(LoginIdentity)														// process signInWithPopup()
-	async loginIdentity(ctx: StateContext<IAuthState>, { authProvider, credential }: LoginIdentity) {
+	private async loginIdentity(ctx: StateContext<IAuthState>, { authProvider, credential }: LoginIdentity) {
 		try {
 			const response = await this.afAuth.auth.signInWithPopup(authProvider);
 
@@ -130,7 +139,7 @@ export class AuthState implements NgxsOnInit {
 
 	/** Attempt to sign in a User via authentication with a Custom Token */
 	@Action(LoginToken)
-	async loginToken(ctx: StateContext<IAuthState>, { token, user, prefix }: LoginToken) {
+	private async loginToken(ctx: StateContext<IAuthState>, { token, user, prefix }: LoginToken) {
 		const additionalUserInfo: firebase.auth.AdditionalUserInfo = {
 			isNewUser: false,
 			providerId: getProviderId(prefix),
@@ -145,7 +154,7 @@ export class AuthState implements NgxsOnInit {
 
 	/** Attempt to sign in a User via an emailAddress / Password combination. */
 	@Action(LoginEmail)															// process signInWithEmailAndPassword
-	loginEmail(ctx: StateContext<IAuthState>, { email, password, method, credential }: LoginEmail) {
+	private loginEmail(ctx: StateContext<IAuthState>, { email, password, method, credential }: LoginEmail) {
 		type TEmailMethod = 'signInWithEmailAndPassword' | 'createUserWithEmailAndPassword';
 		const thisMethod = `${method}WithEmailAndPassword` as TEmailMethod;
 		this.dbg('loginEmail: %s', method);
@@ -169,14 +178,14 @@ export class AuthState implements NgxsOnInit {
 
 	/** Attempt to sign in a User anonymously */
 	@Action(LoginAnon)
-	LoginAnon(ctx: StateContext<IAuthState>) {
+	private loginAnon(ctx: StateContext<IAuthState>) {
 		return this.afAuth.auth.signInAnonymously()
 			.catch(error => ctx.dispatch(new LoginFailed(error)));
 	}
 
 	/** Attempt to sign in a User via a link sent to their emailAddress */
 	@Action(LoginLink)
-	async loginLink(ctx: StateContext<IAuthState>, { link, credential }: LoginLink) {
+	private async loginLink(ctx: StateContext<IAuthState>, { link, credential }: LoginLink) {
 		const localItem = 'emailForSignIn';
 
 		if (this.afAuth.auth.isSignInWithEmailLink(link)) {
@@ -195,25 +204,24 @@ export class AuthState implements NgxsOnInit {
 	}
 
 	@Action(Logout)																	// process signOut()
-	async logout(ctx: StateContext<IAuthState>) {
+	private logout(ctx: StateContext<IAuthState>) {
 		this.user = null;
-		await this.afAuth.auth.signOut();
-
-		ctx.dispatch(new LogoutSuccess());
+		this.afAuth.auth.signOut()
+			.then(_ => ctx.dispatch(new LogoutSuccess()))
 		return;
 	}
 
 	/** Events */
 	@Action(LoginSuccess)														// on each LoginSuccess, fetch /member collection
-	onMember(ctx: StateContext<IAuthState>, { user }: LoginSuccess) {
+	private onMember(ctx: StateContext<IAuthState>, { user }: LoginSuccess) {
 		const query: IQuery = { where: addWhere(FIELD.uid, user.uid) };
 
 		if (this.afAuth.auth.currentUser) {
 			this.sync.on(COLLECTION.attend, query);
 			this.sync.on(COLLECTION.member, query)			// wait for /member snap0 
 				.then(_ => this._memberSubject.next(ctx.getState().info))
-				.then(_ => this.dbg('route: %j', this.navigate.url))
-				.then(_ => {
+				.then(_ => this._memberSubject.complete())
+				.then(_ => {															// if on "/" or "/login", redirect to "/attend"
 					if (['/', '/login'].includes(this.navigate.url))
 						this.navigate.route(ROUTE.attend)
 				})
@@ -221,65 +229,66 @@ export class AuthState implements NgxsOnInit {
 	}
 
 	@Action(AuthInfo)
-	async authInfo(ctx: StateContext<IAuthState>, { info }: AuthInfo) {
+	private async authInfo(ctx: StateContext<IAuthState>, { info }: AuthInfo) {
 		ctx.patchState(info);
 	}
 
 	@Action(AuthOther)															// mimic another User
-	otherMember(ctx: StateContext<IAuthState>, { member }: AuthOther) {
+	private otherMember(ctx: StateContext<IAuthState>, { member }: AuthOther) {
 		const state = ctx.getState();
 
 		if (state.current && state.current.uid === member)
 			return;																			// nothing to do
-		if (state.current && getPath(state.current, 'customClaims.memberName') === member)
+		if (state.current && getPath(state.current, 'customClaims.alias') === member)
 			return;																			// nothing to do
 
 		return this.store.selectOnce<TStateSlice<IRegister>>(state => state[SLICE.admin])
 			.pipe(
 				map(admin => admin[STORE.register]),			// get the Register segment
-				map(table => table.find(row => row[FIELD.uid] === member || getPath(row, 'user.customClaims.memberName') === member)),
+				map(table => table.find(row => row[FIELD.uid] === member || getPath(row, 'user.customClaims.alias') === member)),
 			)
 			.subscribe(reg => ctx.patchState({ current: reg && reg.user }))
 			.unsubscribe()
 	}
 
-	@Action([LoginSetup, LoginSuccess])							// on each LoginSuccess, fetch latest UserInfo, IdToken
-	setUserStateOnSuccess(ctx: StateContext<IAuthState>, { user }: LoginSetup) {
+	@Action([LoginSetup, LoginSuccess])
+	private setUserStateOnSuccess(ctx: StateContext<IAuthState>, { user }: LoginSetup) {
 		if (this.afAuth.auth.currentUser) {
-			this.afAuth.auth.currentUser.getIdTokenResult()
-				.then(token => {
-					ctx.patchState({ user, token, current: user });
-					if (this.hasRole(token, 'admin'))
-						this.sync.on(COLLECTION.admin, undefined,
-							[COLLECTION.member, { where: addWhere(FIELD.store, STORE.status) }]);
-				})
+			const clone = cloneObj(user);								// safe copy of User
+
+			if (!clone.email) {
+				this.snack.error('Must have a valid email address');
+				return ctx.dispatch(new Logout());
+			}
+			ctx.patchState({ user: clone, current: clone });
+			ctx.dispatch(new AuthToken());
 		}
 	}
 
-	private hasRole = (token: firebase.auth.IdTokenResult, role: string) => {
-		return getPath<string[]>({ ...token }, 'claims.claims.roles', [])!.includes(role);
-	}
-
 	@Action(AuthToken)															// fetch latest IdToken
-	setToken(ctx: StateContext<IAuthState>) {
-		let token: firebase.auth.IdTokenResult;
+	private setToken(ctx: StateContext<IAuthState>) {
 		if (this.afAuth.auth.currentUser) {
+			let token: firebase.auth.IdTokenResult;
+			
 			this.afAuth.auth.currentUser.getIdTokenResult(true)
 				.then(result => token = result)
 				.then(_ => ctx.patchState({ token }))
 				.then(_ => this.dbg('customClaims: %j', (ctx.getState().token as firebase.auth.IdTokenResult).claims.claims))
 				.then(_ => {
-					if (this.hasRole(token, 'admin'))
+					const roles = getPath<string[]>({ ...token }, 'claims.claims.roles') || [];
+					if (roles.includes('admin'))
 						this.sync.on(COLLECTION.admin, undefined,
 							[COLLECTION.member, { where: addWhere(FIELD.store, STORE.status) }]);
-					else
+					else {
 						this.sync.off(COLLECTION.admin);
+						this.navigate.route(ROUTE.attend);
+					}
 				})
 		}
 	}
 
 	@Action([LoginFailed, LogoutSuccess])
-	notLogin(ctx: StateContext<IAuthState>, { error }: LoginFailed) {
+	private notLogin(ctx: StateContext<IAuthState>, { error }: LoginFailed) {
 		this.sync.off(COLLECTION.member, true);
 		this.sync.off(COLLECTION.attend, true);
 		this.sync.off(COLLECTION.admin, true);
