@@ -1,17 +1,18 @@
 import { Component, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
 
 import { DataService } from '@dbase/data/data.service';
 import { StateService } from '@dbase/state/state.service';
+import { NavigateService } from '@route/router/navigate.service';
 
 @Component({
 	selector: 'wb-root',
 	templateUrl: './app.component.html',
+	styleUrls: ['./app.component.css'],
 })
 export class AppComponent implements OnInit {
-	public app$ = this.state.getMemberData();								// per-member message
+	public app$ = this.state.getMemberData();
 
-	constructor(readonly data: DataService, readonly state: StateService, private router: Router) { }
+	constructor(readonly data: DataService, readonly state: StateService, readonly navigate: NavigateService) { }
 
 	ngOnInit() { }
 
