@@ -1,31 +1,34 @@
 import { UserInfo, firestore } from 'firebase/app';
 
-import { FIELD, STORE, CONNECT, BONUS } from '@dbase/data/data.define';
+import { FIELD, STORE, TYPE, CONNECT, BONUS, REACT, COLLECTION } from '@dbase/data/data.define';
 import { ISummary } from '@dbase/state/state.define';
-import { TString } from '@lib/type.library';
 import { getSlice } from '@dbase/state/state.library';
 
+import { TString } from '@lib/type.library';
+
 type TStoreConfig = STORE.schema | STORE.config | STORE.default;
-type TStoreClient = STORE.class | STORE.event | STORE.price | STORE.plan | STORE.provider | STORE.schedule | STORE.calendar | STORE.location | STORE.instructor | STORE.bonus | STORE.span | STORE.alert;
-type TStoreUser = STORE.profile | STORE.payment | STORE.gift | STORE.message | STORE.migrate | STORE.attend | STORE.register | STORE.status;
+type TStoreClient = STORE.class | STORE.event | STORE.price | STORE.plan | STORE.provider | STORE.schedule | STORE.calendar | STORE.location | STORE.instructor | STORE.bonus | STORE.span | STORE.alert | STORE.react;
+type TStoreUser = STORE.profile | STORE.payment | STORE.gift | STORE.message | STORE.migrate | STORE.attend | STORE.register | STORE.status | STORE.feedback;
 type TTypeDefault = TStoreClient | 'icon';
 
 type TSpan = 'full' | 'half';
-type TPrice = TSpan | 'topUp' | 'hold' | 'expiry';
-type TPayment = 'topUp' | 'hold' | 'credit' | 'debit';
-type TProfile = 'plan' | 'claim' | 'info' | 'pref';
+type TPayment = TYPE.topUp | 'hold' | 'credit' | 'debit';
+type TProfile = TYPE.plan | TYPE.claim | TYPE.info | TYPE.pref;
 type TStatus = 'account' | 'connect';
 type TSchedule = 'event' | 'class' | 'special';
-type TCalendar = 'event' | 'special'
+type TCalendar = 'event' | 'special';
+
+export type TPrice = TSpan | TYPE.topUp | 'hold' | 'expiry';
 export type TClass = 'AeroStep' | 'HiLo' | 'MultiStep' | 'SingleStep' | 'SmartStep' | 'StepBasic' | 'StepDown' | 'StepIn' | 'Zumba' | 'ZumbaStep';
 export type TPlan = 'member' | 'casual' | 'gratis' | 'student' | 'core' | 'intro';
 export type TProvider = 'identity' | 'oauth' | 'oidc' | 'email' | 'play' | 'phone' | 'anonymous';
 
-export type FBoolean = boolean | undefined | firebase.firestore.FieldValue;
-export type FNumber = number | undefined | firebase.firestore.FieldValue;
-export type FString = TString | undefined | firebase.firestore.FieldValue;
+export type FBoolean = boolean | undefined | firestore.FieldValue;
+export type FNumber = number | undefined | firestore.FieldValue;
+export type FString = TString | undefined | firestore.FieldValue;
 export type FType<T> = Exclude<T, firestore.FieldValue>;
-// These are the meta- and common-fields for a standard collection document
+
+// These are the meta- and common-fields for a standard document
 export interface IMeta {
 	[FIELD.id]: string;										// the id field on the remote database, generated by FireBase
 	[FIELD.create]?: number;							// the time when originally created
@@ -52,9 +55,10 @@ interface IClientBase extends IMeta {
 	[FIELD.key]: string | number;
 	[FIELD.icon]?: string;									// an optional icon for the UI
 }
-interface IUserBase extends IMeta {
+interface IUserBase extends IMeta {				// this is the base for Member-related documents
 	[FIELD.store]: TStoreUser;
 	[FIELD.uid]: string;
+	[FIELD.stamp]: number;
 }
 export interface IMigrateBase extends IMeta {
 	[FIELD.store]: STORE.migrate;
@@ -72,7 +76,7 @@ export interface IStoreMeta extends IMeta {
 export type TStoreMeta = IStoreMeta | IStoreMeta[];
 // client documents have a '<key>' field, user documents have a '<uid>' field
 export const isClientDocument = (document: TStoreBase): document is IClientBase =>
-	getSlice(document[FIELD.store]) === 'client' || getSlice(document[FIELD.store]) === 'local';
+	getSlice(document[FIELD.store]).toString() === 'client' || getSlice(document[FIELD.store]).toString() === 'local';
 
 //	/client/_default_
 export interface IDefault extends IClientBase {
@@ -89,8 +93,8 @@ export interface IConfig extends IClientBase {
 //	/client/_schema_
 export interface ISchema extends IClientBase {
 	[FIELD.store]: STORE.schema;
-	[FIELD.type]: string;
-	[FIELD.key]: string;
+	[FIELD.type]: COLLECTION;
+	[FIELD.key]: STORE;
 	desc: string;
 	filter?: TString;
 	sort?: TString;
@@ -107,12 +111,14 @@ export interface IPrice extends IClientBase {
 //	/client/bonus
 export interface IBonus extends IClientBase {
 	[FIELD.store]: STORE.bonus;
-	[FIELD.key]: 'week' | 'month' | 'sunday' | 'sun3';
+	[FIELD.key]: Exclude<BONUS, BONUS.gift>;	// all Except 'gift'
 	free: number | TString;							// number of free classes available, or string[] of classes available
 	level: number;											// number of classes required to attend previously, before this Bonus
 	rule: TString;
-	desc?: TString;
+	desc: TString;
 	sort?: number;
+	count?: number;
+	amount?: number;										// an optional amount to apply
 }
 
 //	/client/plan
@@ -157,8 +163,16 @@ export interface ICalendar extends IClientBase {
 }
 
 //	/client/schedule
+export interface TBonus {							// a sub-type of IBonus
+	[FIELD.id]: string;
+	[FIELD.type]: BONUS;
+	count: number;											// the number of this Gift
+	desc: TString;											// message to Member
+	gift?: IGift[];											// the /member/gift updates for this payment
+	amount?: number;										// an optional amount (if other than $0)
+}
 export interface ISchedule extends IClientBase {
-	[FIELD.store]: STORE.schedule;
+	[FIELD.store]: STORE.schedule | STORE.calendar;
 	[FIELD.type]: TSchedule;
 	[FIELD.key]: TClass | string;
 	day: number;
@@ -166,8 +180,12 @@ export interface ISchedule extends IClientBase {
 	instructor?: string;
 	start: string;
 	span?: string;
-	price?: number;											// infer the member's price for this class
-	elect?: string;											// name the Bonus the Member chooses
+	price?: IPrice;											// the regular price detail for this class
+	amount?: number;										// infer the member's price for this class
+	bonus?: TBonus;											// the Bonus tracking which can be applied to this Schedule
+	elect?: BONUS;											// name the Bonus the Member chooses (override calc)
+	count?: number;											// number of Attends
+	react?: REACT;											// Member reaction
 }
 
 //	/client/location
@@ -215,9 +233,16 @@ export interface ISpan extends IClientBase {
 //	/client/alert									// <key> is immaterial
 export interface IAlert extends IClientBase {
 	[FIELD.store]: STORE.alert;
-	[FIELD.type]: 'schedule'
+	[FIELD.type]: STORE.schedule | STORE.event;
 	[FIELD.note]: TString;
 	location?: string;
+}
+
+//	/client/react									// Member reaction icons
+export interface IReact extends IClientBase {
+	[FIELD.store]: STORE.react;
+	[FIELD.key]: REACT;
+	sort: number;
 }
 
 //	/client/schedule
@@ -276,19 +301,19 @@ export interface IProfile extends IUserBase {
 	[FIELD.type]: TProfile;
 }
 export interface IProfilePlan extends IProfile {
-	[FIELD.type]: STORE.plan;
+	[FIELD.type]: TYPE.plan;
 	plan: TPlan;
 }
 export interface IProfileClaim extends IProfile {
-	[FIELD.type]: 'claim';
-	claim: ICustomClaims;
+	[FIELD.type]: TYPE.claim;
+	claims: ICustomClaims;
 }
 export interface IProfileInfo extends IProfile {
-	[FIELD.type]: 'info';
-	'info': IMemberInfo;
+	[FIELD.type]: TYPE.info;
+	[TYPE.info]: IMemberInfo;
 }
 export interface IProfilePref extends IProfile {
-	[FIELD.type]: 'pref';
+	[FIELD.type]: TYPE.pref;
 }
 export type TProfileInfo = IProfileInfo | IProfileInfo[];
 
@@ -308,12 +333,16 @@ export interface IMemberInfo {				// Conformed Info across Providers
 export interface IPayment extends IUserBase {
 	[FIELD.store]: STORE.payment;
 	[FIELD.type]: TPayment;
-	[FIELD.stamp]: number;						// create-date
-	amount: number;										// how much actually paid (may be different from plan.topUp.amount)
+	link?: {
+		child?: string;									// the child Payment
+		parent?: string;								// the parent Payment
+		index?: number;									// the Payment position
+	}
+	amount?: number;									// how much actually paid (may be different from plan.topUp.amount)
 	adjust?: number;									// a manual adjustment to a topUp
 	bank?: number;										// un-used funds from previous payment
 	hold?: number;										// number of days to extend expiry
-	expiry?: number;									// six-months from _effect
+	expiry?: number;									// usually six-months from _effect
 	approve?: {												// acknowledge receipt of payment
 		[FIELD.uid]: string;
 		[FIELD.stamp]: number;
@@ -324,30 +353,16 @@ export interface IPayment extends IUserBase {
 //	/member/gift
 export interface IGift extends IUserBase {
 	[FIELD.store]: STORE.gift,
-	// [FIELD.type]: string;
-	[FIELD.stamp]: number;
-	count: number;
+	[FIELD.type]?: string;
+	limit: number;
 	desc?: TString;
 	expiry?: number;
+	count?: number;										// number of Attends so-far for this Gift
 }
 
 // /attend
-export interface TBonus {
-	[FIELD.id]?: string;
-	[FIELD.type]?: BONUS;
-	gift?: IGift[];										// the /member/gift updates for this payment
-	count?: number;										// the number of this Gift
-	scheme?: IBonus;									// the /client/bonus _id for this payment
-}
-interface TTrack {
-	day: number;											// weekDay attended
-	date: number;											// yearMonthDay attended
-	week: number;											// yearWeek attended
-	month: number;										// yearMonth attended
-}
 export interface IAttend extends IUserBase {
 	[FIELD.store]: STORE.attend;
-	[FIELD.stamp]: number;						// the timestamp of the check-in
 	payment: {
 		[FIELD.id]: string;							// the /member/payment _id
 		amount: number;									// the amount the member was charged
@@ -358,22 +373,52 @@ export interface IAttend extends IUserBase {
 		[FIELD.type]: TSchedule;
 		[FIELD.key]: TClass;
 	}
-	bonus?: TBonus;										// override standard price
-	track: TTrack;										// to use in bonus-checking, attend-analysis, etc.
+	bonus?: {
+		[FIELD.id]: string;
+		[FIELD.type]: BONUS;
+		count?: number;									// the count of this Bonus
+	},
+	track: {													// to use in bonus-checking, attend-analysis, etc.
+		date: number;										// yearMonthDay attended
+		day: number;										// weekDay attended (1-7, 1=Mon)
+		week: number;										// yearWeek attended
+		month: number;									// yearMonth attended
+	}
+}
+
+//	/member/feedback
+export interface IFeedBack extends IUserBase {
+	[FIELD.store]: STORE.feedback;
+	[FIELD.type]: STORE.attend;				// currently only 'Attend' feedback
+	[FIELD.key]: string;							// key to document the Member is referencing
+	date: number;											// yyyymmdd of Feedback
+	react?: REACT;
+	comment?: TString;
 }
 
 //	/admin/register
 type TRole = 'admin' | 'member' | 'guest';
 export interface ICustomClaims {		// a special sub-set of fields from the User Token
+	alias?: string;
 	roles?: TRole[];
-	plan?: TPlan;											// if set, we know the Member is fully-defined, otherwise check /member/profile/plan
-	memberName?: string;
-	memberAllow?: string[];
-	memberDeny?: string[];
+	allow?: string[];
+	deny?: string[];
 }
 export type TUser = UserInfo & {
 	customClaims?: ICustomClaims;
 	providerData?: (UserInfo | null)[];
+}
+type TProviderInfo = {
+	id: string;
+	provider: 'fb' | 'g+' | 'gh' | 'tw' | 'li';
+	firstName: string;
+	lastName: string;
+	userName: string;
+	email: string;
+	photoURL?: string;
+	gender?: 'male' | 'female';
+	isAdmin?: true;
+	birthDay?: number;
 }
 export interface IRegister extends IUserBase {
 	[FIELD.store]: STORE.register;
@@ -386,17 +431,7 @@ export interface IRegister extends IUserBase {
 		sheetName: string;
 		uid: string;
 		userName: string;
-		providers: {
-			id: string;
-			provider: 'fb' | 'g+' | 'gh' | 'tw' | 'li';
-			firstName: string;
-			lastName: string;
-			userName: string;
-			email: string;
-			gender?: 'male' | 'female';
-			isAdmin?: boolean;
-			birthDay?: number;
-		}[]
+		providers: TProviderInfo[]
 	};
 }
 
@@ -404,7 +439,6 @@ export interface IRegister extends IUserBase {
 export interface IStatus extends IUserBase {
 	[FIELD.store]: STORE.status;
 	[FIELD.type]: TStatus;
-	stamp: number;
 }
 export interface IStatusAccount extends IStatus {
 	[FIELD.type]: 'account';
