@@ -2,8 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { AngularFireModule } from '@angular/fire';
-import { AngularFirestoreModule, FirestoreSettingsToken } from '@angular/fire/firestore';
-import { AngularFireDatabaseModule } from '@angular/fire/database';
+import { AngularFirestoreModule } from '@angular/fire/firestore';
 import { AngularFireFunctionsModule, FunctionsRegionToken } from '@angular/fire/functions';
 import { environment } from '@env/environment';
 
@@ -23,12 +22,10 @@ const fb = environment.firebase || {};
 		NgxsModule.forFeature([ClientState, MemberState, AttendState, AdminState]),
 		AngularFireModule.initializeApp(fb.app, fb.config),
 		AngularFirestoreModule.enablePersistence({ synchronizeTabs: true }),
-		AngularFireDatabaseModule,
 		AngularFireFunctionsModule,
 	],
 	providers: [
-		{ provide: FunctionsRegionToken, useValue: 'us-central1' },
-		{ provide: FirestoreSettingsToken, useValue: {} },				// TODO: is this still necessary after @angular/fire is upgraded from 5.1.1
+		{ provide: FunctionsRegionToken, useValue: fb.config.region },
 	],
 	declarations: [InfoSnackBar, WarnSnackBar, ErrorSnackBar],
 })
