@@ -2,20 +2,21 @@ import { Observable, defer, combineLatest } from 'rxjs';
 import { switchMap, map } from 'rxjs/operators';
 
 import { TWhere } from '@dbase/fire/fire.interface';
+import { addWhere } from '@dbase/fire/fire.library';
 import { IFireClaims } from '@service/auth/auth.interface';
+import { calcBonus } from '@service/member/attend.library';
 import { getMemberAge } from '@service/member/member.library';
 import { SLICES, SORTBY } from '@library/config.define';
 import { asAt, firstRow, filterTable } from '@library/app.library';
 
-import { IState, IAccountState, ITimetableState, IPlanState, SLICE, TStateSlice, IApplicationState } from '@dbase/state/state.define';
+import { IState, IAccountState, ITimetableState, IPlanState, SLICE, TStateSlice, IApplicationState, ISummary } from '@dbase/state/state.define';
 import { IDefault, IStoreMeta, IClass, IPrice, IEvent, ISchedule, ISpan, IProfilePlan, TStoreBase } from '@dbase/data/data.schema';
-import { STORE, FIELD } from '@dbase/data/data.define';
+import { STORE, FIELD, BONUS, COLLECTION } from '@dbase/data/data.define';
 
 import { asArray } from '@lib/array.library';
 import { getPath, sortKeys, cloneObj, isEmpty } from '@lib/object.library';
 import { isString, isArray, isFunction, isUndefined } from '@lib/type.library';
 import { DATE_FMT, getDate, TDate } from '@lib/date.library';
-import { addWhere } from '@dbase/fire/fire.library';
 
 /**
  * Generic Slice Observable  
@@ -38,7 +39,7 @@ export const getCurrent = <T>(states: IState, store: STORE, filter: TWhere = [],
 /** Get all documents by filter, do not exclude _expire unless <date> specified */
 export const getStore = <T>(states: IState, store: STORE, filter: TWhere = [], date?: TDate) => {
 	const slice = getSlice(store);
-	if (store === slice)													// top-level slice
+	if (store === slice.toString())													// top-level slice
 		return getState<T>(states, store, filter, date);
 
 	const state: Observable<TStateSlice<T>> = states[slice] as any;
@@ -47,8 +48,10 @@ export const getStore = <T>(states: IState, store: STORE, filter: TWhere = [], d
 		throw new Error(`Cannot resolve state from ${store}`);
 
 	return state.pipe(
-		map(obs => isUndefined(date) ? filterTable<T>(obs[store], filter) : asAt<T>(obs[store], filter, date)),
-		map(table => table.sort(sortKeys(...SORTBY[store]))),
+		map(obs => isUndefined(date)
+			? filterTable<T>(obs[store], filter)
+			: asAt<T>(obs[store], filter, date)),
+		map(table => table.sort(sortKeys(...asArray(SORTBY[store])))),
 	)
 }
 export const getState = <T>(states: IState, store: STORE, filter: TWhere = [], date?: TDate) => {
@@ -61,13 +64,13 @@ export const getState = <T>(states: IState, store: STORE, filter: TWhere = [], d
 	return state.pipe(
 		map(obs => Object.keys(obs).map(list => res.concat(Object.values(obs[list]))).flat()),
 		map(table => isUndefined(date) ? filterTable<T>(table, filter) : asAt<T>(table, filter, date)),
-		map(table => table.sort(sortKeys(...SORTBY[store]))),
+		map(table => table.sort(sortKeys(...asArray(SORTBY[store])))),
 	)
 }
 
 export const getSlice = (store: STORE) => {			// determine the state-slice based on the <store> field
 	const slices = Object.keys(SLICES)
-		.filter(col => SLICES[col].includes(store));// find which slice holds the requested store
+		.filter(key => SLICES[key as COLLECTION]!.includes(store));// find which slice holds the requested store
 
 	if (isEmpty<object>(SLICES))									// nothing in State yet, on first-time connect
 		slices.push(SLICE.client);									// special: assume 'client' slice.
@@ -77,12 +80,12 @@ export const getSlice = (store: STORE) => {			// determine the state-slice based
 	if (!slices.length)
 		alert(`Unexpected store: ${store}`)
 
-	return slices[0];
+	return slices[0] as COLLECTION;
 }
 
 export const getDefault = (state: IApplicationState, type: string) => {
 	const table = (state['application'][STORE.default])
-		.filter(row => row[FIELD.type] === type);       // find the default value for the requested type
+		.filter(row => row[FIELD.type] === type);		// find the default value for the requested type
 	return table.length && table[0][FIELD.key] || undefined;
 }
 
@@ -105,10 +108,14 @@ export const joinDoc = (states: IState, node: string | undefined, store: STORE, 
 		return source.pipe(
 			switchMap(data => {
 				const filters = decodeFilter(data, cloneObj(filter)); // loop through filters
-				const segment = (store === STORE.attend) ? filters[0].value : store;	// TODO: dont rely on defined filter
+				// const segment = (store === STORE.attend) ? filters[0].value : store;	// TODO: dont rely on defined filter
 
 				parent = data;                                        // stash the original parent data state
-				return combineLatest(getCurrent<TStoreBase>(states, store, filters, date, segment));
+				// return combineLatest(getCurrent<TStoreBase>(states, store, filters, date, segment));
+				return combineLatest(store === STORE.attend
+					? getStore<TStoreBase>(states, store, filters, date)
+					: getCurrent<TStoreBase>(states, store, filters, date)
+				);
 			}),
 
 			map(res => {
@@ -174,7 +181,7 @@ const decodeFilter = (parent: any, filter: TWhere = []) => {
 
 				return lookup;                              				// rebuild filter's <value>
 			})
-			.distinct()
+			.distinct()																						// remove duplicates
 
 		if (cond.value.length === 1)
 			cond.value = cond.value[0];                           // an array of only one value, return as string
@@ -185,7 +192,7 @@ const decodeFilter = (parent: any, filter: TWhere = []) => {
 
 /**
  * Use the Observable on open IPayment[] to determine current account-status (paid, pending, bank, adjusts, etc.).  
- * The IPayment array is reverse-sorted by 'stamp', meaning payment[0] is the 'active' value.  
+ * The IPayment array is reverse-sorted by 'stamp', meaning payment[0] is the earliest 'active' value.  
  * paid		: is amount from active payment, if topUp
  * adjust : is amount from active payment, if debit  
  * bank		: is any unspent funds (brought forward from previous payment)  
@@ -195,10 +202,12 @@ const decodeFilter = (parent: any, filter: TWhere = []) => {
  */
 export const sumPayment = (source: IAccountState) => {
 	if (source.account) {
+		const sum: ISummary = { paid: 0, bank: 0, adjust: 0, pend: 0, spend: 0, credit: 0, funds: 0 }
+
 		source.account.payment = asArray(source.account.payment);
 		source.account.summary = source.account.payment
 			.reduce((sum, payment, indx) => {
-				if (indx === 0/** && payment[FIELD.effect] */) {		// only 1st Payment
+				if (indx === 0) {																		// only 1st Payment
 					sum.bank += payment.bank || 0;
 					sum.adjust += payment.adjust || 0;
 					sum.paid += payment.amount || 0;
@@ -209,8 +218,7 @@ export const sumPayment = (source: IAccountState) => {
 				sum.funds = sum.bank + sum.paid + sum.adjust;
 
 				return sum;
-			}, { paid: 0, bank: 0, adjust: 0, pend: 0, spend: 0, credit: 0, funds: 0 })
-
+			}, sum)
 	}
 
 	return { ...source };
@@ -225,10 +233,10 @@ export const sumAttend = (source: IAccountState) => {
 				sum.credit -= attend.payment.amount;
 				sum.funds -= attend.payment.amount;
 				return sum;
-			}, source.account.summary);
+			}, source.account.summary)
 	}
 
-	return { ...source };
+	return { ...source }
 }
 
 /** calc a 'day' field for a Calendar row */
@@ -237,7 +245,7 @@ export const calendarDay = (source: ITimetableState) => {
 		source.client.calendar = source.client.calendar
 			.map(row => ({ ...row, day: getDate(row[FIELD.key]).dow }))
 	}
-	return { ...source };
+	return { ...source }
 }
 
 /** Assemble a Plan-view */
@@ -272,14 +280,14 @@ export const buildPlan = (source: IPlanState) => {
 		return plan
 	});
 
-	return { ...source };
+	return { ...source }
 }
 
 /**
  * use the collected Schedule items to determine the Timetable to display.  
  * schedule type 'event' overrides 'class'; 'special' appends.  
  */
-export const buildTimetable = (source: ITimetableState) => {
+export const buildTimetable = (source: ITimetableState, date?: TDate, elect?: BONUS) => {
 	const {
 		schedule: times = [],												// the schedule for the requested date
 		class: classes = [],    										// the classes offered on that date
@@ -287,10 +295,9 @@ export const buildTimetable = (source: ITimetableState) => {
 		event: events = [],													// the event-description for the calendar type
 		span: spans = [],														// the duration of classes / events
 		alert: alerts = [],													// any alert notes for this date
+		price: prices = [],													// the prices as per member's plan
 	} = source.client;
-	const {
-		price: prices = [],
-	} = source.client;														// the prices as per member's plan
+	const attendToday = source.attend.attendToday;
 
 	const icon = firstRow<IDefault>(source.application[STORE.default], addWhere(FIELD.type, 'icon'));
 	const locn = firstRow<IDefault>(source.application[STORE.default], addWhere(FIELD.type, 'location'));
@@ -320,7 +327,6 @@ export const buildTimetable = (source: ITimetableState) => {
 				addWhere(FIELD.type, STORE.event),
 			])
 			const duration = spanClass.duration || span.duration;
-
 			const time: Partial<ISchedule> = {
 				[FIELD.id]: classDoc[FIELD.id],
 				[FIELD.type]: eventList[FIELD.store],
@@ -338,17 +344,23 @@ export const buildTimetable = (source: ITimetableState) => {
 		})
 	})
 
-	// for each item on the schedule, poke in 'price' and 'icon'
-	// TODO: determine bonus-pricing
+	// for each item on the schedule, poke in 'price' and 'icon',
+	// override plan-price if entitled to bonus
 	source.client.schedule = times
 		.map(time => {
 			const classDoc = firstRow<IClass>(classes, addWhere(FIELD.key, time[FIELD.key]));
-			const price = firstRow<IPrice>(prices, addWhere(FIELD.type, classDoc[FIELD.type]));
+			time.bonus = calcBonus(source, classDoc[FIELD.key], date, elect);
+			time.price = firstRow<IPrice>(prices, addWhere(FIELD.type, classDoc[FIELD.type]));
+			time.amount = isUndefined(time.bonus[FIELD.id])	// no Bonus for this class
+				? time.price.amount
+				: (time.bonus.amount || 0)							// a specific-amount, else $0
 
-			if (classDoc[FIELD.type] && !isUndefined(price.amount)) {
-				time.price = time.price || price.amount;	// add-on the member's price for each scheduled event
-			}
-			else time[FIELD.disable] = true;						// cannot determine the event
+			time.count = attendToday.filter(row => row.timetable[FIELD.key] == time[FIELD.key]).length;
+
+			// if (classDoc[FIELD.type] && !isUndefined(price.amount)) {
+			// 	time.price = time.price || price.amount;	// add-on the member's price for each scheduled event
+			// }
+			// else time[FIELD.disable] = true;						// cannot determine the event
 
 			if (!time[FIELD.icon])											// if no schedule-specific icon...
 				time[FIELD.icon] = classDoc.icon || icon[FIELD.key];				//	use class icon, else default icon
@@ -360,133 +372,10 @@ export const buildTimetable = (source: ITimetableState) => {
 		})
 
 		/**
-		 * Special Events take priority over Classes.  
+		 * Special Events take priority over scheduled Classes.  
 		 * remove Classes at Location, if an Event is offered there
 		 */
 		.filter(row => row[FIELD.type] === 'event' || !eventLocations.includes(row.location!))
 
 	return { ...source }
-}
-
-// export const getDash = (source: TStateSlice<IStoreMeta>) => {
-// 	debugger;
-
-// 	const dashBoard = source.register
-// 		.map(register => ({
-// 			register,
-// 			account: source.summary && source.summary.find(acct => acct[FIELD.uid] === register[FIELD.uid]),
-// 		}))
-
-// 	return { ...dashBoard }
-// }
-
-/** Calculate tracking against Bonus schemes */
-export const calcBonus = (event: string, date?: TDate) => {
-
-	// const [uid, gifts, scheme] = await Promise.all([
-	// 	this.data.auth.current
-	// 		.then(user => user!.uid),
-
-	// 	this.data.getStore<IGift>(STORE.gift, undefined, date)					// any active Gifts
-	// 		.then(gifts => asAt(gifts, undefined, date)),									// effective on this date
-
-	// 	this.data.getStore<IBonus>(STORE.bonus, undefined, date)				// any active Bonuses
-	// 		.then(bonus => ({
-	// 			'week': bonus.find(row => row[FIELD.key] === 'week') || {} as IBonus,
-	// 			'month': bonus.find(row => row[FIELD.key] === 'month') || {} as IBonus,
-	// 			'sunday': bonus.find(row => row[FIELD.key] === 'sunday') || {} as IBonus,
-	// 		})),
-	// ]);
-
-	// /**
-	//  * attendGift		is array of Attends so far against 1st active Gift
-	//  * attendWeek		is array of Attends so far in the week, less than <now>
-	//  * attendMonth	is array of Attends so far in the month, less than <now>
-	//  */
-	// const now = getDate(date);
-	// const where = addWhere(FIELD.uid, uid);
-	// const prior = addWhere(`track.${FIELD.date}`, now.format(DATE_FMT.yearMonthDay), '<');
-	// const [attendGift, attendWeek, attendMonth] = await Promise.all([		// get tracking data
-	// 	this.data.getStore<IAttend>(STORE.attend, [where, addWhere(`bonus.${FIELD.id}`, (gifts[0] || {})[FIELD.id])]),
-	// 	this.data.getStore<IAttend>(STORE.attend, [where, addWhere('track.week', now.format(DATE_FMT.yearWeek)), prior]),
-	// 	this.data.getStore<IAttend>(STORE.attend, [where, addWhere('track.month', now.format(DATE_FMT.yearMonth)), prior]),
-	// ]);
-
-	// // We de-dup the Attends by date (yyyymmdd), as multiple attends on a single-day do not count towards a Bonus
-	// let bonus = {} as TBonus;
-	// switch (true) {
-	// 	/**
-	// 	 * Admin adds 'Gift' records to a Member's account which will detail
-	// 	 * the start-date, and count of free classes (and an optional expiry for the Gift).  
-	// 	 * This bonus will be used in preference to any other bonus-pricing scheme.
-	// 	 */
-	// 	case gifts.length > 0:														// qualify for a Gift bonus?
-	// 		bonus = {
-	// 			[FIELD.id]: gifts[0][FIELD.id],
-	// 			[FIELD.type]: 'gift',
-	// 			gift: {
-	// 				[FIELD.expire]: gifts[0].count - attendGift.length <= 1 ? now.ts : undefined,
-	// 				...gifts[0],
-	// 			},
-	// 		}
-	// 		break;
-
-	// 	/**
-	// 	 * The Week scheme qualifies as a Bonus if the Member attends the required number of full-price or gift classes in a week (scheme.week.level).  
-	// 	 * The Member must also have attended less than the free limit (scheme.week.free) to claim this bonus
-	// 	 */
-	// 	case scheme.week
-	// 		&& attendWeek
-	// 			.filter(row => isUndefined(row.bonus) || row.bonus[FIELD.type] === 'gift')
-	// 			.distinct(row => row.track[FIELD.date])
-	// 			.length + 1 > scheme.week.level
-	// 		&& attendWeek
-	// 			.filter(row => row.bonus === scheme.week[FIELD.id])
-	// 			.distinct(row => row.track[FIELD.date])
-	// 			.length < scheme.week.free:
-	// 		bonus = {
-	// 			[FIELD.id]: scheme.week[FIELD.id],
-	// 			[FIELD.type]: 'week',
-	// 		}
-	// 		break;
-
-	// 	/**
-	// 	 * The Sunday scheme qualifies as a Bonus if the Member attends the required number of full-price classes in a week (scheme.sunday.level),
-	// 	 * but does not qualify for the Week scheme (already claimed yesterday?).    
-	// 	 * The class must be in the free list (scheme.week.free) to claim this bonus
-	// 	 */
-	// 	case scheme.sunday
-	// 		&& attendWeek
-	// 			.filter(row => isUndefined(row.bonus) || row.bonus[FIELD.type] === 'gift')
-	// 			.distinct(row => row.track[FIELD.date])
-	// 			.length + 1 > scheme.sunday.level
-	// 		&& now.dow === Instant.DAY.Sun
-	// 		&& (scheme.sunday.free as TString).includes(event):
-	// 		bonus = {
-	// 			[FIELD.id]: scheme.sunday[FIELD.id],
-	// 			[FIELD.type]: 'sunday',
-	// 		}
-	// 		break;
-
-	// 	/**
-	// 	 * The Month scheme qualifies as a Bonus if the Member attends the required number of full-price classes in a month (scheme.month.level).  
-	// 	 * The Member must also have attended less than the free limit (scheme.month.free) to claim this bonus
-	// 	 */
-	// 	case scheme.month
-	// 		&& attendMonth
-	// 			.filter(row => isUndefined(row.bonus) || row.bonus[FIELD.type] === 'gift')
-	// 			.distinct(row => row.track[FIELD.date])
-	// 			.length + 1 > scheme.month.level
-	// 		&& attendMonth
-	// 			.filter(row => row.bonus === scheme.month[FIELD.id])
-	// 			.distinct(row => row.track[FIELD.date])
-	// 			.length < scheme.month.free:
-	// 		bonus = {
-	// 			[FIELD.id]: scheme.month[FIELD.id],
-	// 			[FIELD.type]: 'month',
-	// 		}
-	// 		break;
-	// }
-
-	// return { ...source }
 }
