@@ -2,13 +2,19 @@ import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 
+import { DialogService } from '@service/material/dialog.service';
+import { AttendService } from '@service/member/attend.service';
+
 import { ITimetableState } from '@dbase/state/state.define';
 import { StateService } from '@dbase/state/state.service';
+import { FIELD, COLLECTION } from '@dbase/data/data.define';
+import { IReact } from '@dbase/data/data.schema';
 import { DataService } from '@dbase/data/data.service';
-import { MemberService } from '@service/member/member.service';
 
-import { swipe } from '@lib/html.library';
+import { isUndefined } from '@lib/type.library';
+import { Instant } from '@lib/date.library';
 import { suffix } from '@lib/number.library';
+import { swipe } from '@lib/html.library';
 import { dbg } from '@lib/logger.library';
 
 @Component({
@@ -24,23 +30,49 @@ export class AttendComponent implements OnInit {
 	public timetable$!: Observable<ITimetableState>;
 	public firstPaint = true;                           // indicate first-paint
 
-	constructor(private readonly member: MemberService, public readonly state: StateService, public readonly data: DataService) { }
+	constructor(private readonly attend: AttendService, public readonly state: StateService,
+		public readonly data: DataService, private dialog: DialogService) { }
 
 	ngOnInit() {                                        // wire-up the timetable Observable
 		this.timetable$ = this.state.getScheduleData(this.date).pipe(
 			map(data => {
-				// this.firstPaint = false;                   // TODO: ok to animate if Observable re-emits
 				this.locations = (data.client.location || []).length;
 				this.selectedIndex = 0;                       // start on the first-page
-				// this.dbg('table: %j', data.client);
 				return data;
 			})
 		)
 	}
 
-	// TODO: popup info about a Class
-	showEvent(event: string) {
-		this.dbg('event: %j', event);
+	// Build info to show in a Dialog
+	showEvent(client: ITimetableState["client"], idx: number) {
+		const item = client.schedule![idx];							// the Schedule item clicked
+		const event = client.class!.find(row => row[FIELD.key] === item[FIELD.key]);
+		const locn = client.location!.find(row => row[FIELD.key] === item.location);
+		const instr = client.instructor!.find(row => row[FIELD.key] === item.instructor);
+		const span = client.span!.find(row => row[FIELD.key] === event![FIELD.type]);
+		const dow = Instant.WEEKDAY[item.day];
+		// const bonus = client.bonus!.find(row => row[FIELD.key] === item.bonus);
+
+		const title = item.key;
+		const subtitle = event && event.desc || '';
+		const icon = item.icon;
+		const actions = ['Close'];
+
+		let content: string[] = [];
+		if (!isUndefined(item.amount))
+			content.push(`Cost: $${item.amount.toFixed(2)}`);
+		if (item.start)
+			content.push(`Start: ${dow} @ ${item.start}`)
+		if (span)
+			content.push(`Duration: ${span.duration} minutes`);
+		if (item.count)
+			content.push(`Attended: ${item.count} times today`);
+		if (locn)
+			content.push(`Location: ${locn.name}`);
+		if (instr)
+			content.push(`Instructor: ${instr.name}`);
+
+		this.dialog.open({ content, title, subtitle, icon, actions });
 	}
 
 	swipe(idx: number, event: any) {
