@@ -7,7 +7,6 @@ import { NavigateService } from '@route/navigate.service';
 import { LoginModule } from '@route/login/login.module';
 
 import { AuthModule } from '@service/auth/auth.module';
-import { AuthService } from '@service/auth/auth.service';
 import { addWhere } from '@dbase/fire/fire.library';
 import { LState } from '@dbase/state/state.define';
 import { StoreStorage } from '@dbase/sync/sync.define';
@@ -15,25 +14,24 @@ import { FIELD } from '@dbase/data/data.define';
 import { IProfilePlan } from '@dbase/data/data.schema';
 import { asAt } from '@library/app.library';
 
-import { getLocalStore } from '@lib/window.library';
-import { isUndefined } from '@lib/type.library';
+import { getLocalStore } from '@lib/browser.library';
 import { getPath } from '@lib/object.library';
 import { dbg } from '@lib/logger.library';
 
+/**
+ * TODO: The ROUTE.oauth should not be accessible by the Member.
+ * It is provided solely so that an OAuth provider has a redirect back to the App
+ * (after the Member authenticates).  
+ */
 @Injectable({ providedIn: AuthModule })
-export class AuthGuard implements CanActivate {
+export class OAuthGuard implements CanActivate {
 	private dbg = dbg(this);
 
-	constructor(private auth: AuthService, private navigate: NavigateService) { this.dbg('new') }
+	constructor(private navigate: NavigateService) { this.dbg('new') }
 
 	async canActivate() {
-		const state = await this.auth.user;
-
-		if (state.auth.user)												// is logged-on
-			return true;
-
-		this.navigate.route(ROUTE.login);						// redirect to LoginComponent
-		return false;
+		this.dbg('oauth: %s', this.navigate.url);
+		return true;
 	}
 }
 
@@ -42,7 +40,7 @@ export class AuthGuard implements CanActivate {
 export class ProfileGuard implements CanActivate {
 	private dbg = dbg(this);
 
-	constructor(private auth: AuthService, private navigate: NavigateService) { this.dbg('new') }
+	constructor(private navigate: NavigateService) { this.dbg('new') }
 
 	async canActivate() {
 		const localState = getLocalStore<LState>(StoreStorage) || {};
@@ -51,10 +49,10 @@ export class ProfileGuard implements CanActivate {
 		if (getPath<string>(planProfile, 'plan'))
 			return true;															// found a current 'plan' in localStorage
 
-		const state = await this.auth.user;
-		const planClaim = getPath<string>(state.auth, 'token.claims.claims.plan');
-		if (!isUndefined(planClaim))
-			return true;															// found a 'plan' on customClaims token
+		// const state = await this.auth.user;
+		// const planClaim = getPath<string>(state.auth, 'token.claims.claims.plan');
+		// if (!isUndefined(planClaim))
+		// 	return true;															// found a 'plan' on customClaims token
 
 		this.navigate.route(ROUTE.plan);						// redirect to PlanComponent
 		return false;
