@@ -8,6 +8,7 @@ import { FIELD, STORE } from '@dbase/data/data.define';
 import { IStoreMeta, IGift } from '@dbase/data/data.schema';
 
 import { getStamp, TDate, getDate } from '@lib/date.library';
+import { TString } from '@lib/type.library';
 import { asArray } from '@lib/array.library';
 import { dbg } from '@lib/logger.library';
 
@@ -47,7 +48,7 @@ export class AdminService {
 	}
 
 	/** Create a Gift for a Member to use in future check-ins */
-	addGift = async (uid: string, count: number, type?: string, start?: TDate, expiry?: TDate) => {
+	addGift = async (uid: string, limit: number, type?: string, start?: TDate, note?: TString, expiry?: TDate) => {
 		return this.data.setDoc(STORE.gift, {
 			[FIELD.effect]: getDate(start).startOf('day').ts,
 			[FIELD.store]: STORE.gift,
@@ -55,8 +56,8 @@ export class AdminService {
 			[FIELD.uid]: uid,
 			stamp: getStamp(),
 			expiry: expiry && getStamp(expiry),
-			count: count,
+			limit: limit,
+			note: note,
 		} as IGift)
 	}
-
 }
