@@ -1,30 +1,66 @@
 import { Component, Injectable, Inject } from '@angular/core';
 import { MatDialog, MatDialogRef, MatDialogConfig, MAT_DIALOG_DATA } from '@angular/material';
 
+import { dbg } from '@lib/logger.library';
+import { asArray } from '@lib/array.library';
+import { TString } from '@lib/type.library';
+
 @Component({
   selector: 'info-dialog',
-  template: '<span style="color:green">Note: </span>{{ data }}'
+  template: `
+    <div mat-dialog-title>
+      <img [src]="data.icon || 'https://source.unsplash.com/random/200x200'">
+      <h2> {{ data.title }} </h2>
+    </div>
+
+    <mat-dialog-content class="mat-typography">
+      <h3> {{ data.subtitle }} </h3><p>
+      <div *ngFor="let content of data.content">
+        {{ content }}
+      </div>
+    </mat-dialog-content>
+
+    <mat-dialog-actions align="end">
+      <div *ngFor="let btn of data.actions">
+        <button mat-button mat-dialog-close>{{ btn }}</button>
+      </div>
+    </mat-dialog-actions>
+  `,
 })
 export class InfoDialogComponent {
-  constructor(@Inject(MAT_DIALOG_DATA) public data: any) { }
+  constructor(private dialogRef: MatDialogRef<InfoDialogComponent>, @Inject(MAT_DIALOG_DATA) public data: any, ) { }
+
+  close() {
+    this.dialogRef.close();
+  }
+}
+
+interface openDialog {
+  content: TString;
+  icon?: string;
+  title?: string;
+  subtitle?: string;
+  actions?: string[];
 }
 
 @Injectable({ providedIn: 'root' })
 export class DialogService {
-	private ref?: MatDialogRef<{}>;
+  private ref?: MatDialogRef<{}>;
+  private dbg = dbg(this);
 
   constructor(private dialog: MatDialog) { }
 
-  public open(msg: string, action?: string, config: MatDialogConfig = {}) {
-		this.dismiss();
-		this.ref = this.dialog.open(InfoDialogComponent, config);
-    // this.ref = this.snack.open(msg, action, this.setConfig(config));
+  public open(data: openDialog, config: MatDialogConfig = {}) {
+    config.data = data;
+
+    this.dismiss();
+    this.ref = this.dialog.open(InfoDialogComponent, config);
   }
-	
-	public dismiss() {                    // dismiss any snackbar, if present
+
+  public dismiss() {                    // dismiss any MatDialog, if present
     if (this.ref) {
-      // this.ref.dismiss();
+      this.ref.close();
       this.ref = undefined;
-		}
-	}
+    }
+  }
 }
