@@ -1,4 +1,6 @@
 import { Injectable } from '@angular/core';
+import { combineLatest } from 'rxjs';
+import { take } from 'rxjs/operators';
 
 import { addWhere } from '@dbase/fire/fire.library';
 import { StateService } from '@dbase/state/state.service';
@@ -6,194 +8,57 @@ import { sumPayment, sumAttend } from '@dbase/state/state.library';
 import { SyncAttend } from '@dbase/state/state.action';
 import { STORE, FIELD, BONUS } from '@dbase/data/data.define';
 
+import { PAY, ATTEND } from '@service/member/attend.define';
 import { calcExpiry } from '@service/member/member.library';
 import { MemberService } from '@service/member/member.service';
 import { SnackService } from '@service/material/snack.service';
 import { DBaseModule } from '@dbase/dbase.module';
 import { TWhere } from '@dbase/fire/fire.interface';
 import { DataService } from '@dbase/data/data.service';
-import { IAttend, IStoreMeta, TClass, TStoreBase, ISchedule, IPayment, IGift, IBonus, TBonus } from '@dbase/data/data.schema';
+import { IAttend, IStoreMeta, TClass, TStoreBase, ISchedule, IPayment, IGift } from '@dbase/data/data.schema';
 
-import { getDate, DATE_FMT, TDate, Instant } from '@lib/date.library';
+import { getDate, DATE_FMT, TDate } from '@lib/date.library';
 import { isUndefined, isNumber, TString } from '@lib/type.library';
-import { getPath, isEmpty, IObject } from '@lib/object.library';
+import { getPath, isEmpty, sortKeys } from '@lib/object.library';
 import { asArray } from '@lib/array.library';
 import { dbg } from '@lib/logger.library';
-import { randomString } from '@lib/string.library';
 
+/** Manage the 'attend' store */
 @Injectable({ providedIn: DBaseModule })
 export class AttendService {
 	private dbg = dbg(this);
 
 	constructor(private member: MemberService, private state: StateService, private data: DataService, private snack: SnackService) { this.dbg('new'); }
 
-	/** look back up-to-seven days to find when className was last scheduled */
-	lkpDate = async (className: string, location?: string, date?: TDate) => {
+	/** look back up-to seven days to find when className was last scheduled */
+	private lkpDate = async (className: string, location?: string, date?: TDate) => {
 		const timetable = await this.state.getTimetableData(date).toPromise();
-		let now = getDate(date);													// start with date-argument
+		let now = getDate(date);																// start with date-argument
 		let ctr = 0;
 
 		if (!location)
 			location = this.state.getDefault(timetable, STORE.location);
 
 		for (ctr; ctr < 7; ctr++) {
-			const classes = timetable.client.schedule!			// loop through schedule
-				.filter(row => row.day === now.dow)						// finding a match in 'day'
-				// .filter(row => row.start === now.format(DATE_FMT.HHMI)),	// TODO: match by time, in case offered multiple times
-				.filter(row => row[FIELD.key] === className)	// and match in 'class'
-				.filter(row => row.location === location)			// and match in 'location'
+			const classes = timetable.client.schedule!						// loop through schedule
+				.filter(row => row.day === now.dow)									// finding a match in 'day'
+				.filter(row => row[FIELD.key] === className)				// and match in 'class'
+				.filter(row => row.location === location)						// and match in 'location'
+				// .filter(row => row.start === now.format(DATE_FMT.HHMI)),	// TODO: match by time, in case offered multiple times in a day
 
-			if (classes.length)															// is this class offered on this 'day'   
+			if (classes.length)																		// is this class offered on this 'day'   
 				break;
-			now = now.add(-1, 'day');												// move pointer to previous day
+			now = now.add(-1, 'day');															// else move pointer to previous day
 		}
 
-		if (ctr >= 7)																			// cannot find className on timetable
-			now = getDate(date);														// so default back to today's date	
+		if (ctr >= 7)																						// cannot find className on timetable
+			now = getDate(date);																	// so default back to today's date	
 
-		return now.ts;																		// timestamp
+		return now.ts;																					// timestamp
 	}
 
-	/**
-	 * Determine the eligibility for free Attend
-	 */
-	async getBonus(event: string, date?: TDate, elect?: string) {
-		/**
-		 * uid is the Member to check-in 
-		 * gifts is array of active Gifts effective (not expired) on this date  
-		 * scheme is definition of active Bonus schemes effective on this date
-		 */
-		const [uid, gifts, scheme] = await Promise.all([
-			this.data.auth.current
-				.then(user => user!.uid),
-
-			this.data.getStore<IGift>(STORE.gift, undefined, date),					// any active Gifts
-
-			this.data.getStore<IBonus>(STORE.bonus, undefined, date)				// any active Bonuses
-				.then(bonus => bonus.reduce((acc, row) => {
-					acc[row[FIELD.key]] = row;
-					return acc;
-				}, {} as IObject<IBonus>))
-		]);
-
-		/**
-		 * attendGift		is array of Attends[] against each active Gifts
-		 * attendWeek		is array of Attends so far in the week, less than <now>
-		 * attendMonth	is array of Attends so far in the month, less than <now>
-		 */
-		const now = getDate(date);
-		const where = addWhere(FIELD.uid, uid);
-		const prior = addWhere(`track.${FIELD.date}`, now.format(DATE_FMT.yearMonthDay), '<');
-		const [attendGift, attendWeek, attendMonth] = await Promise.all([		// get tracking data
-			Promise.all(																			// return an array of Attend per Gift
-				gifts.map(gift => this.data.getStore<IAttend>(STORE.attend, [where, addWhere(`bonus.${FIELD.id}`, gift[FIELD.id])]))
-			),
-			this.data.getStore<IAttend>(STORE.attend, [where, addWhere('track.week', now.format(DATE_FMT.yearWeek)), prior]),
-			this.data.getStore<IAttend>(STORE.attend, [where, addWhere('track.month', now.format(DATE_FMT.yearMonth)), prior]),
-		]);
-
-		// We de-dup the Attends by date (yyyymmdd), as multiple attends on a single-day do not count towards a Bonus
-		let bonus = {} as TBonus;
-		const upd: IGift[] = [];														// an array of updates to Gifts
-		switch (true) {
-			/**
-			 * Admin adds 'Gift' records to a Member's account which will detail
-			 * the start-date, and count of free classes (and an optional expiry for the Gift).  
-			 * This bonus will be used in preference to any other bonus-pricing scheme.
-			 */
-			case gifts.length > 0:														// Member has some open Gifts
-				let curr = -1;																	// the Gift to use in determining eligibility
-				gifts.forEach((gift, idx) => {
-					if (attendGift[idx].length >= gifts[idx].count// max number of Attends against this gift
-						|| (gifts[idx].expiry || 0) > now.ts)				// 	or this Gift expiry has passed
-						upd.push({ [FIELD.expire]: now.ts, ...gift })// auto-expire it
-					else if (curr === -1) curr = idx;							// else found a Gift that might apply
-				})
-
-				if (curr !== -1) {
-					const attendCnt = attendGift[curr].length;		// how many Attends already against this Gift
-					const gift = gifts[curr];											// this is the Gift to book against
-					bonus = {
-						[FIELD.id]: gift[FIELD.id],									// Bonus id is the Gift Id
-						[FIELD.type]: BONUS.gift,										// Bonus type is 'gift'
-						count: attendCnt + 1,												// to help with tracking
-						gift: [...upd, {
-							[FIELD.effect]: gift[FIELD.effect] || now.startOf('day').ts,
-							[FIELD.expire]: gift.count - attendCnt <= 1 ? now.ts : undefined,
-							...gift,
-						}],
-					}
-					break;																				// exit switch-case
-				}																								// drop through to next switch-case
-
-			/**
-			 * The Week scheme qualifies as a Bonus if the Member attends the required number of full-price or gift classes in a week (scheme.week.level).  
-			 * The Member must also have attended less than the free limit (scheme.week.free) to claim this bonus
-			 */
-			case scheme.week
-				&& attendWeek
-					.filter(row => isUndefined(row.bonus) || row.bonus[FIELD.type] === BONUS.gift)
-					.distinct(row => row.track[FIELD.date])
-					.length + 1 > scheme.week.level
-				&& attendWeek
-					.filter(row => row.bonus === scheme.week[FIELD.id])
-					.distinct(row => row.track[FIELD.date])
-					.length < scheme.week.free:
-				bonus = {
-					[FIELD.id]: scheme.week[FIELD.id],
-					[FIELD.type]: BONUS.week,
-					gift: upd,
-				}
-				break;
-
-			/**
-			 * The Sunday scheme qualifies as a Bonus if the Member attends the required number of full-price classes in a week (scheme.sunday.level),
-			 * and does not qualify for the Week scheme (already claimed yesterday?).    
-			 * The class must be in the free list (scheme.week.free) to claim this bonus
-			 */
-			case scheme.sunday
-				&& attendWeek																				// count Attends this week either Gift or non-Bonus
-					.filter(row => isUndefined(row.bonus) || row.bonus[FIELD.type] === BONUS.gift)
-					// .distinct(row => row.track[FIELD.date])				// de-dup by day-of-week
-					.length + 1 > scheme.sunday.level									// required number of Attends this week
-				&& now.dow === Instant.DAY.Sun											// today is 'Sunday'
-				&& elect == BONUS.sunday														// Member elected to take Bonus
-				&& (scheme.sunday.free as TString).includes(event):	// Event is offered on Bonus
-				bonus = {
-					[FIELD.id]: scheme.sunday[FIELD.id],
-					[FIELD.type]: BONUS.sunday,
-					gift: upd,
-				}
-				break;
-
-			/**
-			 * The Month scheme qualifies as a Bonus if the Member attends the required number of full-price classes in a month (scheme.month.level).  
-			 * The Member must also have attended less than the free limit (scheme.month.free) to claim this bonus
-			 */
-			case scheme.month
-				&& attendMonth
-					.filter(row => isUndefined(row.bonus) || row.bonus[FIELD.type] === BONUS.gift)
-					.distinct(row => row.track[FIELD.date])
-					.length + 1 > scheme.month.level
-				&& attendMonth
-					.filter(row => row.bonus === scheme.month[FIELD.id])
-					.distinct(row => row.track[FIELD.date])
-					.length < scheme.month.free:
-				bonus = {
-					[FIELD.id]: scheme.month[FIELD.id],
-					[FIELD.type]: BONUS.month,
-					gift: upd,
-				}
-				break;
-		}
-
-		return bonus;
-	}
-
-	/**
-	 * Insert an Attend document, matched to an active Payment  
-	 */
-	async setAttend(schedule: ISchedule, note?: TString, date?: TDate) {
+	/** Insert an Attend document, aligned to an active Payment  */
+	public setAttend = async (schedule: ISchedule, note?: TString, date?: TDate) => {
 		const creates: IStoreMeta[] = [];
 		const updates: IStoreMeta[] = [];
 
@@ -206,10 +71,27 @@ export class AttendService {
 		const stamp = now.ts;
 		const when = now.format(DATE_FMT.yearMonthDay);
 
-		let data = await this.member.getAccount(date);		// get Member's current account details
+		// There is a bit of overlap in these Observables, but they serve different purposes
+		let [data, source] = await combineLatest(
+			this.state.getAccountData(date),								// get Member's current account details
+			this.state.getScheduleData(date, schedule.elect),// get Schedule for this date
+		)
+			.pipe(take(1))																	// unsubscribe on first-emit
+			.toPromise()																		// emit observables as Promise
+
+																											// <schedule> might be from Schedule or Calender store
+		const field = schedule[FIELD.store] === STORE.schedule
+			? FIELD.id																			// compare requested schedule to timetable's ID
+			: FIELD.key																			// compare requested event to class's Key
+		const timetable = source.client.schedule!.find(row => row[field] === schedule[field]);
+		if (!timetable) {																	// this schedule is not actually offered on this date!
+			this.dbg(`Cannot find this schedule item: ${schedule[FIELD.id]}`);
+			this.snack.error(`Cannot find this schedule item: ${schedule[FIELD.id]}`);
+			return false;																		// this should not happen
+		}
 
 		// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-		// check we are not re-booking same Class on same Day at same Location at same Time
+		// check we are not re-booking same Class on same Day in same Location at same Time
 		const attendFilter = [
 			addWhere(`track.${FIELD.date}`, when),
 			addWhere(`timetable.${FIELD.key}`, schedule[FIELD.key]),
@@ -226,33 +108,27 @@ export class AttendService {
 
 		// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 		// work out the actual price to charge
-		let [bonus, calcPrice] = await Promise.all([
-			data.client.plan[0].bonus
-				? this.getBonus(schedule[FIELD.key], date, schedule.elect)// any bonus entitlement
-				: {} as TBonus,
-			this.member.getEventPrice(schedule[FIELD.key], data),
-		]);
-
-		if (!isEmpty(bonus)) {
-			if (bonus.gift) {
-				updates.push(...bonus.gift);									// batch the Gift updates
-				delete bonus.gift;														// not needed
+		if (!isEmpty(timetable.bonus) && timetable.bonus) {
+			if (timetable.bonus.gift) {
+				updates.push(...timetable.bonus.gift);						// batch any Gift updates
+				delete timetable.bonus.gift;											// not needed
+				delete timetable.bonus.desc;											// not needed
 			}
-			if (bonus[FIELD.id])
-				calcPrice = 0;																// override calculated price
+			if (timetable.bonus[FIELD.id])
+				timetable.price!.amount = timetable.bonus.amount || 0;	// override Plan price
 		}
 
-		if (!isUndefined(schedule.price) && schedule.price !== calcPrice) {// calculation mis-match
-			this.dbg('bonus: %j', bonus);
-			this.snack.error(`Price discrepancy: paid ${schedule.price}, but should be ${calcPrice}`);
-			throw new Error(`Price discrepancy: paid ${schedule.price}, but should be ${calcPrice}`);
+		// if the caller provided a schedule.amount, and it is different to the calculated amount!
+		if (!isUndefined(schedule.amount) && schedule.amount !== timetable.price!.amount) {// calculation mis-match
+			this.dbg('bonus: %j', timetable.bonus);
+			this.snack.error(`Price discrepancy: paid ${schedule.amount}, but should be ${timetable.price!.amount}`);
+			throw new Error(`Price discrepancy: paid ${schedule.amount}, but should be ${timetable.price!.amount}`);
 		}
-		schedule.price = calcPrice;
+		schedule.amount = timetable.price!.amount;				// the price to use for this class
 
 		// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-		// loop through Payments to determine which is <active>
-		const tests = new Set<boolean>();									// Set of test results
-		let active: IPayment;															// Payment[0]
+		// loop through Payments to determine which should be <active>
+		let active: IPayment;															// Payment[0] is assumed active, by default
 
 		while (true) {
 			data = sumPayment(data);												// calc Payment summary
@@ -260,26 +136,38 @@ export class AttendService {
 			active = data.account.payment[0];								// current Payment
 			const expiry = getPath<number>(active, 'expiry') || Number.MAX_SAFE_INTEGER;
 
-			tests.add(data.account.attend.length < 100);						// arbitrary limit of Attends against a Payment             
-			tests.add(schedule.price <= data.account.summary.funds);// enough funds to cover this Attend
-			tests.add(now.ts < expiry);															// Payment expired
-			if (!tests.has(false))
-				break;																				// all tests passed
-			tests.clear();																	// reset the test Set
-			this.dbg('limit: %j, %j', data.account.attend.length < 100, data.account.attend.length);
-			this.dbg('funds: %j, %j, %j', schedule.price <= data.account.summary.funds, schedule.price, data.account.summary);
-			this.dbg('expiry: %j, %j, %j', now.ts < expiry, now.ts, expiry);
+			const tests: boolean[] = [];										// Set of test results
+			tests[PAY.under_limit] = data.account.attend.length < ATTEND.maxColumn;
+			tests[PAY.enough_funds] = schedule.amount <= data.account.summary.funds;
+			tests[PAY.not_expired] = now.ts <= expiry;
+			if (!tests.includes(false))											// if tests do not include at-least one <false>
+				break;																				// 	all tests passed; we have found a Payment
 
-			if (data.account.payment.length <= 1) {					// create a new Payment
-				const gap = schedule.price - data.account.summary.credit;
-				const topUp = await this.member.getPayPrice(data);
-				const payment = await this.member.setPayment(Math.max(gap, topUp));
+			this.dbg('limit: %j, %j', tests[PAY.under_limit], data.account.attend.length);
+			this.dbg('funds: %j, %j, %j', tests[PAY.enough_funds], schedule.amount, data.account.summary);
+			this.dbg('expiry: %j, %j, %j', tests[PAY.not_expired], now.ts, expiry);
 
-				if (topUp === 0)
-					payment.approve = { uid: 'Auto-Approve', stamp: now.ts }
+			// If over-limit (100+ Attends per Payment) but still have sufficient funds to cover this Class price,
+			// 	insert an auto-approved $0 Payment (only if the next Attend is earlier than any future Payments)
+			if (!tests[PAY.under_limit] && tests[PAY.enough_funds] /** && tests[PAY.not_expired] */) {
+				if (data.account.payment.length <= 1 || stamp < data.account.payment[1].stamp) {
+					const payment = await this.member.setPayment(0, stamp);
+					payment.approve = { uid: ATTEND.autoApprove, stamp };
+
+					creates.push(payment);											// stack the topUpPayment into the batch
+					data.account.payment.splice(1, 0, payment);	// make this the 'next' Payment
+				}
+			}
+
+			// If there are no future Payments, create a new Payment with the higher of
+			// a. the usual topUp amount, or b. the amount needed to cover this Attend
+			if (data.account.payment.length <= 1) {
+				const gap = schedule.amount - data.account.summary.credit;
+				const topUp = Math.max(gap, await this.member.getPayPrice(data));
+				const payment = await this.member.setPayment(topUp, stamp);
 
 				creates.push(payment);												// stack the topUp Payment into the batch
-				data.account.payment.push(payment);						// append to the Payment array
+				data.account.payment.push(payment);						// make this the 'next' Payment
 			}
 
 			const next = data.account.payment[1];
@@ -291,21 +179,29 @@ export class AttendService {
 				data.account.payment = data.account.payment.slice(1);// drop the 1st, now-inactive payment
 				data.account.attend = await this.data.getStore(STORE.attend, addWhere(`${STORE.payment}.${FIELD.id}`, next[FIELD.id]));
 			}
-		}
+
+			if (isUndefined(next)) {
+				this.snack.error('Could not allocate a new Payment');
+				throw new Error('Could not allocate a new Payment');
+			}
+		}																									// repeat the loop to test if this now-Active Payment is useable
 
 		const upd: Partial<IPayment> = {};								// updates to the Payment
 		if (!active[FIELD.effect])
 			upd[FIELD.effect] = stamp;											// mark Effective on first check-in
-		if (data.account.summary.funds === schedule.price && schedule.price)
+		if (data.account.summary.funds === schedule.amount && schedule.price)
 			upd[FIELD.expire] = stamp;											// mark Expired if no funds (except $0 classes)
-		if (!active.expiry && active.approve)							// calc an Expiry for this Payment
+		if (!active.expiry && active.approve && isEmpty(timetable.bonus))	// calc an Expiry for this Payment
 			upd.expiry = calcExpiry(active.approve.stamp, active, data.client);
 
-		if (Object.keys(upd).length)
-			updates.push({ ...active, ...upd });						// batch the Payment update
+		if (Object.keys(upd).length)											// changes to the active Payment
+			updates.push({ ...active, ...upd });						// so, batch the Payment update
 
+		if (data.member.plan[0].plan === 'intro' && (data.account.summary.funds - schedule.amount) <= 0)
+			this.member.setPlan('member', stamp + 1);				// auto-bump 'intro' to 'member' Plan
 		// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 		// got everything we need; write an Attend document
+
 		const attendDoc: Partial<IAttend> = {
 			[FIELD.store]: STORE.attend,
 			[FIELD.stamp]: stamp,														// createDate
@@ -313,12 +209,12 @@ export class AttendService {
 			timetable: {
 				[FIELD.id]: schedule[FIELD.id],								// <id> of the Schedule
 				[FIELD.key]: schedule[FIELD.key] as TClass,		// the Attend's class
-				[FIELD.store]: schedule[FIELD.type] === 'class' ? 'schedule' : 'calendar',
+				[FIELD.store]: schedule[FIELD.type] === 'class' ? STORE.schedule : STORE.calendar,
 				[FIELD.type]: schedule[FIELD.type],						// the type of Attend ('class','event','special')
 			},
 			payment: {
 				[FIELD.id]: active[FIELD.id],									// <id> of Account's current active document
-				amount: schedule.price,												// calculated price of the Attend
+				amount: schedule.amount,											// calculated price of the Attend
 			},
 			track: {
 				[FIELD.date]: when,														// yyyymmdd of the Attend
@@ -326,14 +222,15 @@ export class AttendService {
 				week: now.format(DATE_FMT.yearWeek),					// week-number of year
 				month: now.format(DATE_FMT.yearMonth),				// month-number of year
 			},
-			bonus: !isEmpty(bonus) ? bonus : undefined,			// <id>/<type> of Bonus
+			bonus: !isEmpty(timetable.bonus) ? timetable.bonus : undefined,			// <id>/<type>/<count> of Bonus
 		}
 		creates.push(attendDoc as TStoreBase);						// batch the new Attend
 
-		data.account.attend.push(attendDoc as IAttend);
-		await this.member.setAccount(creates, updates, data);
+		// data.account.attend.push(attendDoc as IAttend);
+		// await this.member.setAccount(creates, updates, data);
 
 		return this.data.batch(creates, updates, undefined, SyncAttend)
+			.then(_ => this.member.setAccount(creates, updates, data))
 	}
 
 	/**
@@ -342,50 +239,101 @@ export class AttendService {
 	 * Take great care when deleting a non-latest Attend, as this will affect Bonus pricing, Gift tracking, Bank rollovers, etc.
 	 */
 	public delAttend = async (where: TWhere) => {
-		const memberUid = addWhere(FIELD.uid, (await this.data.auth.current)!.uid);
+		const memberUid = addWhere(FIELD.uid, (await this.data.getUID()));
 		const filter = asArray(where);
 		if (!filter.map(clause => clause.fieldPath).includes(FIELD.uid))
 			filter.push(memberUid);
 
-		const [attends, payments, gifts] = await Promise.all([
-			this.data.getStore<IAttend>(STORE.attend, filter),
-			this.data.getStore<IPayment>(STORE.payment, [memberUid, addWhere(FIELD.store, STORE.payment)]),
-			this.data.getStore<IGift>(STORE.gift, [memberUid, addWhere(FIELD.store, STORE.gift)]),
-		])
-		const creates: IStoreMeta[] = [];
 		const updates: IStoreMeta[] = [];
-		const deletes: IStoreMeta[] = [...attends];
+		const deletes: IStoreMeta[] = await this.data.getStore<IAttend>(STORE.attend, filter);
+
+		if (deletes.length === 0) {
+			this.dbg('No items to delete');
+			return;
+		}
+
+		const payIds = deletes												// the Payment IDs related to this reversal
+			.map(attend => attend.payment[FIELD.id])
+			.distinct()
+		const giftIds = deletes												// the Gift IDs related to this reversal
+			.filter(attend => attend.bonus && attend.bonus[FIELD.type] === BONUS.gift)
+			.map(row => row.bonus[FIELD.id])
+			.distinct()
+
+		/**
+		 * attends:		all the Attends that relate to the payIds (even those about to be deleted)
+		 * payments:	all the Payments related to this Member
+		 * gifts:			all the Gifts related to this reversal
+		 */
+		const [attends, payments, gifts] = await Promise.all([
+			this.data.getStore<IAttend>(STORE.attend, addWhere(`payment.${FIELD.id}`, payIds)),
+			this.data.getStore<IPayment>(STORE.payment, memberUid),
+			this.data.getStore<IGift>(STORE.gift, [memberUid, addWhere(FIELD.id, giftIds)]),
+		])
+
+		// build a link-link of Payments, assume pre-sorted by stamp [desc]
+		payments
+			.sort(sortKeys('-' + FIELD.stamp, FIELD.type))
+			.forEach((payment, idx, arr) => {
+				payment.link = {
+					parent: idx + 1 < arr.length && arr[idx + 1][FIELD.id] || undefined,
+					child: idx !== 0 && arr[idx - 1][FIELD.id] || undefined,
+					index: idx,
+				}
+			})
 
 		let pays = attends														// count the number of Attends per Payment
-			.distinct(row => row.payment[FIELD.id])
+			.map(row => row.payment[FIELD.id])
 			.reduce((acc, itm) => {
-				acc[itm] = payments.filter(row => row[FIELD.id] === itm).length;
+				if (isUndefined(acc[itm]))
+					acc[itm] = 0;
+				acc[itm] += 1;
 				return acc;
-			}, {} as IObject<number>);
+			}, {} as Record<string, number>);
 
-		attends.forEach(attend => {
+		const chg = new Set<string>();
+		deletes.forEach(attend => {
+			let idx: number;
 			if (attend.bonus && attend.bonus[FIELD.type] === BONUS.gift) {
-				const idx = gifts.findIndex(row => row[FIELD.id] === attend.bonus![FIELD.id]!);
-				if (idx >= 0 && isNumber(gifts[idx][FIELD.expire])) {
-					gifts[idx][FIELD.expire] = undefined;
-					updates.push({ ...gifts[idx] });				// re-open a Gift
+				idx = gifts.findIndex(row => row[FIELD.id] === attend.bonus![FIELD.id]);
+				const gift = gifts[idx];
+				if (gift) {
+					if (isNumber(gift.count))
+						gift.count! -= 1;											// decrement running-total
+					if (gift.count === 0)
+						gift.count = undefined;								// discard running-total
+					if (isNumber(gift[FIELD.expire]))
+						gift[FIELD.expire] = undefined;				// re-open a Gift
 				}
 			}
 
 			const payId = attend.payment[FIELD.id];			// get the Payment id
-			pays[payId] -= 1;														// reduce the number of Attends for this Payment
-			if (pays[payId] === 0) {
-				const idx = payments.findIndex(row => row[FIELD.id] === payId);
-				if (idx > 0) {														// re-open a Payment
-					payments[idx][FIELD.expire] = undefined;
-					payments[idx].bank = undefined;
-					updates.push({ ...payments[idx] })
+			idx = payments.findIndex(row => row[FIELD.id] === payId);
+			const payment = payments[idx];
+			if (payment) {
+				if (isNumber(payment[FIELD.expire]))
+					payment[FIELD.expire] = undefined;			// re-open a Payment
+				pays[payId] -= 1;													// reduce the number of Attends for this Payment
+				if (pays[payId] === 0) {									// no other Attend booked against this Payment
+					payment[FIELD.effect] = undefined;
+					payment.bank = undefined;								// TODO: dont clear down Payment, if parent had $0 funds
+
+					chg.add(payId);													// mark this Payment as 'changed
+					const parent = payment.link!.parent;
+					if (parent) {
+						payments[idx + 1][FIELD.expire] = undefined;
+						chg.add(payments[idx + 1][FIELD.id]);	// mark the parent as 'changed'
+					}
 				}
 			}
-
 		})
+		updates.push(...gifts);												// batch the updates to Gifts
+		updates.push(...payments
+			.filter(row => chg.has(row[FIELD.id]))			// only those that have changed
+			.map(row => ({ ...row, link: undefined }))	// drop the calculated link
+		);
 
-		return this.data.batch(creates, updates, deletes)
+		return this.data.batch(undefined, updates, deletes)
 			.then(_ => this.member.updAccount())
 	}
 }
