@@ -4,9 +4,10 @@ import { IPlanState } from '@dbase/state/state.define';
 import { FIELD } from '@dbase/data/data.define';
 import { IProfileInfo, IMemberInfo, IPayment } from '@dbase/data/data.schema';
 
-import { asArray } from '@lib/array.library';
-import { getStamp, getDate } from '@lib/date.library';
 import { isString, isObject, isNumber, isUndefined } from '@lib/type.library';
+import { asArray } from '@lib/array.library';
+import { cloneObj } from '@lib/object.library';
+import { getStamp, getDate } from '@lib/date.library';
 
 // Library of member-related functions
 
@@ -44,31 +45,40 @@ export const getMemberInfo = (provider: firebase.auth.AdditionalUserInfo) => {
 	if (profileInfo.birthDay)															// as number
 		profileInfo.birthDay = getStamp(profileInfo.birthDay)
 
-	return profileInfo;
+	return cloneObj(profileInfo);													// removed 'undefined' fields
 }
 
-// each Provider might report a different birthday; take latest
+// each Provider might report a different birthday; take earliest
 export const getMemberBirthDay = (info: IProfileInfo[] = []) => {
 	const birthDays = asArray(info)
 		.map(row => row.info.birthDay)
 		.filter(isNumber)
-	return birthDays.length ? Math.max(...birthDays) : undefined;
+	return birthDays.length ? Math.min(...birthDays) : undefined;
 }
 
 export const getMemberAge = (info: IProfileInfo[] = []) =>
 	getDate(getMemberBirthDay(info)).diff('years');				// diff from today
 
-/** A Member's payment will auto-expire (i.e. unused funds lapse) after a number of months */
+/**
+ * A Member's payment will auto-expire (i.e. unused funds lapse) if not fully-used after a number of months.  
+ * 1. Expiry will not be set until a Payment becomes active (all earlier Payments expired).
+ * 2. When the nightly batch expires a prior Payment, it uses the Payment.stamp as the 'from' date.
+ * 3. When the first Attend against a Payment is made, it uses the Attend.stamp to reset the 'from' date.  
+ */
 export const calcExpiry = (stamp: number, payment: IPayment, client: IPlanState["client"]) => {
-	const plan = client.plan[0] || {};								// description of Member's current Plan
+	const plan = client.plan[0] || {};										// description of Member's current Plan
 	const topUp = client.price.find(row => row[FIELD.type] === 'topUp');
 	const hold = payment.hold || 0;
+	const paid = (payment.amount || 0) + (payment.adjust || 0) + (payment.bank || 0);
 
 	if (topUp && !isUndefined(plan.expiry)) {							// plan.expiry is usually six-months
-		const offset = topUp.amount
-			? Math.round(payment.amount / (topUp.amount / plan.expiry)) || 1
+		const offset = topUp.amount													// number of months to extend expiry-date
+			? Math.round(paid / (topUp.amount / plan.expiry)) || 1
 			: plan.expiry;																		// allow for gratis account expiry
-		return getDate(stamp).add(offset, 'months').add(hold, 'days').startOf('day').ts;
+		return getDate(stamp)
+			.add(offset, 'months')														// number of months to extend expiry-date
+			.add(hold, 'days')																// number of days to *hold* off expiry
+			.startOf('day').ts																// set to beginning of day
 	}
 
 	return undefined;
