@@ -13,4 +13,4 @@ if (environment.production)
 
 platformBrowserDynamic()
   .bootstrapModule(AppModule)
-  .catch(err => console.log(err))
+  .catch(err => console.error(err));
